feat(deposit): put deposit enter amount screen into storybook and make it work better
