+++ packages/blockchain-wallet-v4-frontend/src/modals/Brokerage/Banks/Deposit/BankList/template.success.tsx
@@ -2,17 +2,14 @@ import React, { ReactElement } from 'react'
 import { FormattedMessage } from 'react-intl'
 import styled from 'styled-components'
 
+import { BeneficiaryType, NabuSymbolNumberType } from '@core/types'
 import { Icon, Image, Text } from 'blockchain-info-components'
-import {
-  BeneficiaryType,
-  NabuSymbolNumberType
-} from '@core/types'
 import { AddNewButton } from 'components/Brokerage'
 import { FlyoutWrapper } from 'components/Flyout'
+import { Bank, BankWire } from 'components/Flyout/model'
 import { BankDWStepType, BankTransferAccountType } from 'data/types'
 import { getBankLogoImageName } from 'services/images'
 
-import { Bank, BankWire } from '../../model'
 import { Props as _P } from '.'
 
 type OwnProps = {
@@ -87,6 +84,7 @@ const BankList = (props: Props) => {
         ))}
         {props.beneficiaries.map((beneficiary) => (
           <BankWire
+            key={beneficiary.id}
             beneficiary={beneficiary}
             minAmount={props.minAmount}
             onClick={() => {
