+++ packages/blockchain-wallet-v4-frontend/src/modals/Brokerage/Banks/Deposit/EnterAmount/index.tsx
@@ -4,17 +4,24 @@ import { bindActionCreators, Dispatch } from 'redux'
 
 import { Remote } from '@core'
 import { SBPaymentMethodType } from '@core/network/api/simpleBuy/types'
-import { ExtractSuccess, RemoteDataType, WalletFiatType } from '@core/types'
-import { FlyoutOopsError } from 'components/Flyout'
+import { ExtractSuccess, RemoteDataType, SBPaymentTypes, WalletFiatType } from '@core/types'
+import { EnterAmount, FlyoutOopsError } from 'components/Flyout'
+import { getDefaultMethod } from 'components/Flyout/model'
+import { maximumAmount } from 'components/Flyout/validation'
 import { actions, selectors } from 'data'
 import { RootState } from 'data/rootReducer'
-import { BankDWStepType, BankPartners, BankTransferAccountType } from 'data/types'
+import {
+  AddBankStepType,
+  BankDWStepType,
+  BankPartners,
+  BankTransferAccountType,
+  BrokerageModalOriginType
+} from 'data/types'
 
 import { Loading, LoadingTextEnum } from '../../../../components'
 import getData from './selectors'
-import Success from './template.success'
 
-const EnterAmount = (props) => {
+const EnterAmountContainer = (props) => {
   useEffect(() => {
     if (props.fiatCurrency && !Remote.Success.is(props.data)) {
       props.buySellActions.fetchPaymentMethods(props.fiatCurrency)
@@ -46,6 +53,30 @@ const EnterAmount = (props) => {
     })
   }, [])
 
+  const handleBack = useCallback(
+    () =>
+      props.brokerageActions.setDWStep({
+        dwStep: BankDWStepType.DEPOSIT_METHODS
+      }),
+    []
+  )
+
+  const handleChangeMethod = useCallback(() => {
+    props.brokerageActions.setDWStep({
+      dwStep: BankDWStepType.BANK_LIST
+    })
+  }, [])
+
+  const handleAddMethod = useCallback(() => {
+    props.brokerageActions.showModal({
+      modalType: props.fiatCurrency === 'USD' ? 'ADD_BANK_YODLEE_MODAL' : 'ADD_BANK_YAPILY_MODAL',
+      origin: BrokerageModalOriginType.ADD_BANK_DEPOSIT
+    })
+    props.brokerageActions.setAddBankStep({
+      addBankStep: AddBankStepType.ADD_BANK
+    })
+  }, [props.fiatCurrency])
+
   return props.data.cata({
     Failure: () => (
       <FlyoutOopsError
@@ -56,14 +87,41 @@ const EnterAmount = (props) => {
     ),
     Loading: () => <Loading text={LoadingTextEnum.LOADING} />,
     NotAsked: () => <Loading text={LoadingTextEnum.LOADING} />,
-    Success: (val) => (
-      <Success
-        {...val}
-        {...props}
-        onSubmit={onSubmit}
-        initialValues={{ currency: props.fiatCurrency }}
-      />
-    )
+    Success: (val) => {
+      const isUserEligible =
+        val.paymentMethods.methods.length &&
+        val.paymentMethods.methods.find((method) => method.limits.max !== '0')
+      const paymentAccount = getDefaultMethod(props.defaultMethod, val.bankTransferAccounts)
+      const paymentMethod = val.paymentMethods.methods.find(
+        (method) => method.type === SBPaymentTypes.BANK_TRANSFER
+      )
+      let handleMethodClick: () => void
+
+      if (val.bankTransferAccounts.length > 0) {
+        handleMethodClick = handleChangeMethod
+      } else {
+        handleMethodClick = handleAddMethod
+      }
+
+      return isUserEligible ? (
+        <EnterAmount
+          asyncValidate={maximumAmount(paymentMethod.limits.max)}
+          onSubmit={onSubmit}
+          initialValues={{ currency: props.fiatCurrency }}
+          fiatCurrency={props.fiatCurrency}
+          handleBack={handleBack}
+          handleMethodClick={handleMethodClick}
+          paymentAccount={paymentAccount}
+          paymentMethod={paymentMethod}
+        />
+      ) : (
+        <FlyoutOopsError
+          action='retry'
+          data-e2e='sbTryCurrencySelectionAgain'
+          handler={errorCallback}
+        />
+      )
+    }
   })
 }
 
@@ -97,4 +155,4 @@ export type LinkStatePropsType = {
 export type Props = OwnProps & ConnectedProps<typeof connector>
 export type ValidateProps = Props & SuccessStateType & LinkStatePropsType
 
-export default connector(EnterAmount)
+export default connector(EnterAmountContainer)
