+++ packages/blockchain-wallet-v4-frontend/src/modals/Brokerage/Banks/Withdraw/EnterAmount/template.success.tsx
@@ -4,18 +4,18 @@ import { isEmpty } from 'ramda'
 import { Field, InjectedFormProps, reduxForm } from 'redux-form'
 import styled from 'styled-components'
 
-import { Button, Icon, Text } from 'blockchain-info-components'
 import Currencies from '@core/exchange/currencies'
 import { BeneficiaryType, NabuSymbolNumberType } from '@core/types'
+import { Button, Icon, Text } from 'blockchain-info-components'
 import { BlueCartridge, ErrorCartridge } from 'components/Cartridge'
 import CoinDisplay from 'components/Display/CoinDisplay'
 import { AmountTextBox } from 'components/Exchange'
 import { FlyoutWrapper } from 'components/Flyout'
+import { DepositOrWithdrawal, normalizeAmount } from 'components/Flyout/model'
 import { Form } from 'components/Form'
 import { BankTransferAccountType, UserDataType, WithdrawCheckoutFormValuesType } from 'data/types'
 
 import { Row } from '../../components'
-import { DepositOrWithdrawal, normalizeAmount } from '../../model'
 import { Props as OwnProps, SuccessStateType } from '.'
 import Beneficiary from './Beneficiary'
 import LockTimeTooltip from './LockTimeTooltip'
