+++ packages/blockchain-wallet-v4-frontend/src/modals/RecurringBuys/Summary/index.tsx
@@ -5,10 +5,11 @@ import moment from 'moment'
 import { bindActionCreators, Dispatch } from 'redux'
 import styled from 'styled-components'
 
-import { Button, Icon, Text } from 'blockchain-info-components'
 import { fiatToString } from '@core/exchange/utils'
 import { FiatType, SBOrderType, WithdrawalLockCheckRule } from '@core/types'
-import { FlyoutContainer, FlyoutContent, FlyoutFooter, getPeriodTitleText } from 'components/Flyout'
+import { Button, Icon, Text } from 'blockchain-info-components'
+import { FlyoutContainer, FlyoutContent, FlyoutFooter } from 'components/Flyout'
+import { getPeriodTitleText } from 'components/Flyout/model'
 import { actions, selectors } from 'data'
 import { getCounterAmount, getCounterCurrency } from 'data/components/simpleBuy/model'
 import { RootState } from 'data/rootReducer'
