+++ packages/blockchain-wallet-v4-frontend/src/modals/SimpleBuy/EnterAmount/Checkout/template.success.tsx
@@ -9,7 +9,8 @@ import { coinToString, fiatToString } from '@core/exchange/utils'
 import { CoinType, OrderType, SBPaymentMethodType, SBPaymentTypes } from '@core/types'
 import { Banner, Icon, Text } from 'blockchain-info-components'
 import { AmountTextBox } from 'components/Exchange'
-import { FlyoutOopsError, FlyoutWrapper, getPeriodTitleText } from 'components/Flyout'
+import { FlyoutOopsError, FlyoutWrapper } from 'components/Flyout'
+import { getPeriodTitleText } from 'components/Flyout/model'
 import { Form } from 'components/Form'
 import { model } from 'data'
 import { convertStandardToBase } from 'data/components/exchange/services'
